+++ src/app/(main)/chat/(workspace)/_layout/Desktop/Portal.tsx
@@ -39,8 +39,9 @@ const PortalPanel = memo(({ children }: PropsWithChildren) => {
   const { styles } = useStyles();
   const { md = true } = useResponsive();
 
-  const [showInspector, showToolUI] = useChatStore((s) => [
+  const [showInspector, showToolUI, showArtifactUI] = useChatStore((s) => [
     chatPortalSelectors.showPortal(s),
+    chatPortalSelectors.showPluginUI(s),
     chatPortalSelectors.showArtifactUI(s),
   ]);
 
@@ -54,7 +55,7 @@ const PortalPanel = memo(({ children }: PropsWithChildren) => {
         expand
         hanlderStyle={{ display: 'none' }}
         maxWidth={MAX_WIDTH}
-        minWidth={showToolUI ? CHAT_DOCK_TOOL_UI_WIDTH : CHAT_DOCK_WIDTH}
+        minWidth={showArtifactUI || showToolUI ? CHAT_DOCK_TOOL_UI_WIDTH : CHAT_DOCK_WIDTH}
         mode={md ? 'fixed' : 'float'}
         placement={'right'}
         showHandlerWhenUnexpand={false}