+++ src/app/(main)/discover/_layout/Mobile/index.tsx
@@ -1,0 +1,16 @@
+import { PropsWithChildren } from 'react';
+
+import NProgress from '@/components/NProgress';
+
+const Layout = ({ children }: PropsWithChildren) => {
+  return (
+    <>
+      <NProgress />
+      {children}
+    </>
+  );
+};
+
+Layout.displayName = 'MobileDiscoverStoreLayout';
+
+export default Layout;