+++ src/app/(main)/settings/llm/ProviderList/providers.tsx
@@ -1,6 +1,7 @@
 import { useMemo } from 'react';
 
 import {
+  Ai21ProviderCard,
   Ai360ProviderCard,
   AnthropicProviderCard,
   BaichuanProviderCard,
@@ -28,6 +29,7 @@ import {
 import { ProviderItem } from '../type';
 import { useAzureProvider } from './Azure';
 import { useBedrockProvider } from './Bedrock';
+import { useGithubProvider } from './Github';
 import { useOllamaProvider } from './Ollama';
 import { useOpenAIProvider } from './OpenAI';
 
@@ -36,6 +38,7 @@ export const useProviderList = (): ProviderItem[] => {
   const OllamaProvider = useOllamaProvider();
   const OpenAIProvider = useOpenAIProvider();
   const BedrockProvider = useBedrockProvider();
+  const GithubProvider = useGithubProvider();
 
   return useMemo(
     () => [
@@ -47,12 +50,14 @@ export const useProviderList = (): ProviderItem[] => {
       GoogleProviderCard,
       DeepSeekProviderCard,
       OpenRouterProviderCard,
-      GroqProviderCard,
+      GithubProvider,
       NovitaProviderCard,
-      PerplexityProviderCard,
-      MistralProviderCard,
       TogetherAIProviderCard,
       FireworksAIProviderCard,
+      GroqProviderCard,
+      PerplexityProviderCard,
+      MistralProviderCard,
+      Ai21ProviderCard,
       UpstageProviderCard,
       QwenProviderCard,
       SparkProviderCard,
@@ -66,6 +71,6 @@ export const useProviderList = (): ProviderItem[] => {
       SiliconCloudProviderCard,
       TaichuProviderCard,
     ],
-    [AzureProvider, OllamaProvider, OpenAIProvider, BedrockProvider],
+    [AzureProvider, OllamaProvider, OpenAIProvider, BedrockProvider, GithubProvider],
   );
 };