+++ src/app/api/plugin/store/route.ts
@@ -1,10 +1,7 @@
 import { DEFAULT_LANG } from '@/const/locale';
-
-import { PluginStore } from './Store';
+import { PluginStore } from '@/server/modules/PluginStore';
 
 export const runtime = 'edge';
-
-export const revalidate = 43_200; // revalidate at almost every 12 hours
 
 export const GET = async (req: Request) => {
   const locale = new URL(req.url).searchParams.get('locale');